pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama-proxy"
version = "0.1.0"
description = "Unified OpenAI-compatible API router for Ollama and vLLM backends"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Dennis Reichenberg" }]
keywords = ["ollama", "vllm", "openai", "proxy", "router", "llm", "api", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "fastapi>=0.111",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "uvicorn[standard]>=0.29",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "httpx>=0.27",
    "anyio[trio]",
    "ruff",
    "mypy",
]

[project.scripts]
ollama-proxy = "ollama_proxy.cli:main"

[project.urls]
Homepage = "https://github.com/dennisreichenberg/weiterentwicklung"
Repository = "https://github.com/dennisreichenberg/weiterentwicklung"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["ollama_proxy"]
omit = ["*/tests/*"]
